sistema-de-pedidos-spring-boot-2 Create pedidos page

+++ src/main/java/org/ufpr/sistemapedidos/repository/ProdutoRepository.java
@@ -1,6 +1,5 @@
 package org.ufpr.sistemapedidos.repository;
 
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -9,7 +8,7 @@ import org.ufpr.sistemapedidos.model.Produto;
 import java.util.List;
 
 @Repository
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 
     @Query("SELECT c FROM Produto c WHERE c.id = ?1")
     Produto findOne(Integer produtoId);
